automated_trading_system/requirements.txt
python-okx~=0.3.4
mplfinance~=0.12.10b0
pandas~=2.2.3
joblib~=1.4.2
matplotlib~=3.9.2
numpy~=2.0.1
ta~=0.11.0